+++ CHANGELOG.md
@@ -9,6 +9,12 @@ You can also check the [release page](https://github.com/visualize-admin/visuali
 
 ## Unreleased
 
+Nothing yet.
+
+# [3.24.2] - 2023-11-28
+
+- Features
+  - Implemented initial version of merging the cubes (not yet exposed through UI)
 - Fixes
   - Conslidated behavior of setting initial filters (top-most hierarchy value) when filter was not present and multi-filter was removed
   - Fixed switching between segmentation dimensions in column charts